Changed return type of getOwnerId() to 'Long' to match previous change

+++ autoregistry/src/main/java/com/kelley/autoregistry/model/Owner.java
@@ -54,7 +54,7 @@ public class Owner {
 
     // Getters and Setters (No setter for ownerId - Uses @Id annotation - Hibernate will access field directly)
 
-    public Integer getOwnerId() {
+    public Long getOwnerId() {
         return ownerId;
     }
